requirements.txt
flet
python-dotenv
google-generativeai
anthropic
requests
supabase
pydantic
python-dotenv
python-multipart
uvicorn
